PSY-424: split contracts/interfaces.go into per-domain files
Move ~40 service interfaces from the catch-all contracts/interfaces.go
into the existing per-domain files (catalog.go, engagement.go, admin.go,
auth.go, user.go, notification.go, pipeline.go, label.go, festival.go,
release.go, charts.go, radio.go) and delete interfaces.go entirely.

Pure mechanical move within the contracts package — same interface
definitions, same package, no signature changes. Each interface now
lives alongside its associated request/response types.

Also relocate AuditLog{Filters,Response} from notification.go to
admin.go where they belong by domain (admin concern, not notification).

Trim services/interfaces.go to only include the two services that live
in the root services package (CollectionService, RequestService). The
catalog compile-time checks move to a new services/catalog/interfaces.go,
matching the per-sub-package convention used by admin/auth/engagement/
notification/pipeline/user.
