+++ backend/internal/services/contracts/festival.go
@@ -147,3 +147,26 @@ type AddFestivalVenueRequest struct {
 	VenueID   uint `json:"venue_id" validate:"required"`
 	IsPrimary bool `json:"is_primary"`
 }
+
+// ──────────────────────────────────────────────
+// Festival Service Interface
+// ──────────────────────────────────────────────
+
+// FestivalServiceInterface defines the contract for festival operations.
+type FestivalServiceInterface interface {
+	CreateFestival(req *CreateFestivalRequest) (*FestivalDetailResponse, error)
+	GetFestival(festivalID uint) (*FestivalDetailResponse, error)
+	GetFestivalBySlug(slug string) (*FestivalDetailResponse, error)
+	ListFestivals(filters map[string]interface{}) ([]*FestivalListResponse, error)
+	SearchFestivals(query string) ([]*FestivalListResponse, error)
+	UpdateFestival(festivalID uint, req *UpdateFestivalRequest) (*FestivalDetailResponse, error)
+	DeleteFestival(festivalID uint) error
+	GetFestivalArtists(festivalID uint, dayDate *string) ([]*FestivalArtistResponse, error)
+	AddFestivalArtist(festivalID uint, req *AddFestivalArtistRequest) (*FestivalArtistResponse, error)
+	UpdateFestivalArtist(festivalID, artistID uint, req *UpdateFestivalArtistRequest) (*FestivalArtistResponse, error)
+	RemoveFestivalArtist(festivalID, artistID uint) error
+	GetFestivalVenues(festivalID uint) ([]*FestivalVenueResponse, error)
+	AddFestivalVenue(festivalID uint, req *AddFestivalVenueRequest) (*FestivalVenueResponse, error)
+	RemoveFestivalVenue(festivalID, venueID uint) error
+	GetFestivalsForArtist(artistID uint) ([]*ArtistFestivalListResponse, error)
+}
